modified order, stack order

+++ push_swap/include/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: myko <myko@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 14:01:14 by myko              #+#    #+#             */
-/*   Updated: 2022/11/17 17:04:41 by myko             ###   ########.fr       */
+/*   Updated: 2022/11/17 17:52:39 by myko             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@
 typedef struct s_stack {
 	long long	a_len;
 	long long	b_len;
-	long long	*a;
-	long long	*b;
+	int			*a;
+	int			*b;
 }	t_stack;
 
 typedef enum s_stack_num {
@@ -33,9 +33,9 @@ typedef enum s_stack_num {
 long long	ft_atoi_int(const char *str);
 
 // order.c
-void		s_order(long long *stack, long long len);
+void		s_order(int *stack, long long len);
 void		p_order(t_stack *stack, long long goal);
-void		r_order(long long *stack, long long len);
-void		rr_order(long long *stack, long long len);
+void		r_order(int *stack, long long len);
+void		rr_order(int *stack, long long len);
 
 #endif
